+++ lib/active_sanity/invalid_record.rb
@@ -1,6 +1,7 @@
 class InvalidRecord < ActiveRecord::Base
-  belongs_to :record, :polymorphic => true
+  belongs_to :record, polymorphic: true
   serialize :validation_errors
 
-  validates_presence_of :record, :validation_errors
+  validates :record, :validation_errors, presence: true
+  validates :record_id, presence: true, uniqueness: { scope: :record_type }
 end